+++ src/components/forms/contributor/create_category.tsx
@@ -5,7 +5,7 @@ import React, { useState } from "react";
 import { trpc } from "../../../utils/trpc";
 
 import FormTemplate from '../main';
-import { AlertForm } from '../../alert';
+import { AlertForm } from '../../utils/alert';
 import { type Category } from "@prisma/client";
 import { type CategoriesWithChildren } from "../../types";
